requirements.txt
pandas>=2.0
openpyxl>=3.1
matplotlib>=3.7
reportlab>=4.0
pyyaml>=6.0